fix(host_info): split into it's own files
