+++ bundle.list
@@ -54,8 +54,18 @@
 ./code_cert.prod.lua
 ./crashreport.lua
 ./endpoint.lua
-./host_info.lua
 ./init.lua
+./hostinfo/base.lua
+./hostinfo/cpu.lua
+./hostinfo/disk.lua
+./hostinfo/filesystem.lua
+./hostinfo/init.lua
+./hostinfo/memory.lua
+./hostinfo/network.lua
+./hostinfo/nil.lua
+./hostinfo/procs.lua
+./hostinfo/system.lua
+./hostinfo/who.lua
 ./lua_modules/async/init.lua
 ./lua_modules/async/queue.lua
 ./lua_modules/bourbon/init.lua
